% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Private.R
\name{VariantFinder_SurvivalKM_Single}
\alias{VariantFinder_SurvivalKM_Single}
\title{VariantFinder_Survival_Single}
\usage{
VariantFinder_SurvivalKM_Single(
  patient,
  trt,
  variant.ann,
  SNP,
  FilterBy = NULL,
  legend = "Legend",
  labels = NULL,
  xlim = NULL,
  main = "Kaplan-Meier Survival Curve",
  ylab = "Probability of survival",
  xlab = "Time since diagnosis (days)",
  color = NULL,
  height = 5,
  width = 12,
  dpi = 300,
  pvalue = TRUE,
  conf.range = TRUE
)
}
\arguments{
\item{patient}{is the data.frame of the patient clinical data downloaded from MMRF-Commpass Researcher Gateway 
(i.e. MMRF_CoMMpass_IA15_PER_PATIENT file) and imported into environment.}

\item{trt}{is the data.frame of the patient clinical data (i.e. treatment-response) downloaded from MMRF-Commpass Researcher Gateway 
(i.e. MMRF_CoMMpass_IA15_STAND_ALONE_TRTRESP file) and imported into environment.}

\item{variant.ann}{is the data.frame of the annotated variants downloaded from MMRF-Commpass Researcher Gateway 
(i.e. MMRF_CoMMpass_IA15a_All_Canonical_Variants file) and imported into environment.}

\item{legend}{Legend title of the figure}

\item{labels}{labels of the plot}

\item{xlim}{x axis limits e.g. xlim = c(0, 1000). Present narrower X axis, but not affect survival estimates.}

\item{main}{main title of the plot}

\item{ylab}{y axis text of the plot}

\item{xlab}{x axis text of the plot}

\item{color}{Define the colors/Pallete for lines.}

\item{height}{Image height}

\item{width}{Image width}

\item{dpi}{Figure quality}

\item{pvalue}{show p-value of log-rank test}

\item{conf.range}{show confidence intervals for point estimates of survival curves.}

\item{Listvariant}{is the list of the variants to analyze.}

\item{filename}{The name of the pdf file.}
}
\value{
Survival plot
}
\description{
Creates a survival plot from MMRF-RG patient clinical data
using survival library. It uses the fields D_PT_deathdy, D_PT_PRIMARYREASON and D_PT_lstalive
columns for groups.
}
